feat: temp
클라이언트 병합 임시 커밋

+++ Client/Source/Blaster/GameInstance/BlasterGameInstance.h
@@ -78,11 +78,18 @@ public:
 	void SetPlayerId(int32 NewPlayerId) { PlayerId = NewPlayerId; }
 
 private:
+	//UPROPERTY(Config)
+	//FString DefaultServerIP = TEXT("147.185.221.23");
+
+	//UPROPERTY(Config)
+	//int32 DefaultServerPort = 63753;
+
 	UPROPERTY(Config)
 	FString DefaultServerIP = TEXT("127.0.0.1");
 
 	UPROPERTY(Config)
-	int16 DefaultServerPort = 7777;
+	int32 DefaultServerPort = 7777;
+
 
 	UPROPERTY()
 	class UBlasterNetworkSubsystem* NetworkSystem;
